using dtype

+++ python/llm/src/bigdl/llm/transformers/gguf/models/mistral.py
@@ -77,7 +77,7 @@ def load_gguf_mistral(loader: GGUFFileLoader, dtype: torch.dtype = torch.float):
         model = MistralForCausalLM(mistral_config)
 
     for name, weight in state_dict.items():
-        set_module_tensor_to_device(model, name, "cpu", weight)
+        set_module_tensor_to_device(model, name, "cpu", weight, dtype=dtype)
 
     model = model.cpu()
